pyproject.toml
[project]
name = "semantic-kg"
version = "0.1.0"
description = ""
authors = [
    {name = "Morris"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "poethepoet",
    "opencv-python (>=4.11.0.86,<5.0.0.0)",
    "nibabel (>=5.3.2,<6.0.0)",
    "numpy (>=2.2.5,<3.0.0)",
    "torchsummary (>=1.5.1,<2.0.0)",
    "nilearn (>=0.11.1,<0.12.0)",
    "matplotlib (>=3.10.3,<4.0.0)",
    "scikit-image (>=0.25.2,<0.26.0)",
    "langgraph (>=0.4.5,<0.5.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "grad-cam (>=1.5.5,<2.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "google-adk (>=1.3.0,<2.0.0)",
    "google-cloud-aiplatform (>=1.97.0,<2.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
]

[tool.poe.tasks]
install-ltt = "python3 -m pip install light-the-torch"
run-ltt = "python3 -m light_the_torch install --upgrade torch torchaudio torchvision"
autoinstall-torch-cuda = ["install-ltt", "run-ltt"]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
